pyproject.toml
[project]
name = "coursebox"
version = "0.1.0"
description = ""
authors = [
    {name = "Max_Smotr",email = "smotrichenko@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=12.1.0,<13.0.0)",
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "coverage (>=7.13.3,<8.0.0)",
    "drf-yasg (>=1.21.14,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.1,<8.0.0)",
    "eventlet (>=0.40.4,<0.41.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "black (>=26.1.0,<27.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]
